Complete 1512

+++ Problems/dynamic_array/q_1512_number_of_good_pairs.py
@@ -13,6 +13,23 @@ from typing import List
 #         return ans_count
 
 # Solution #2
+# sum(List) => return int or float
+#
+# myList = [1,1,2,3,4,5,3,2,3,4,2,1,2,3]
+# print collections.Counter(myList)
+# Counter({2: 4, 3: 4, 1: 3, 4: 2, 5: 1})
+#
+# print collections.Counter(myList).values()
+# [3, 4, 4, 2, 1]
+#
+# samenumber_count  pairs
+# 0                 0
+# 1                 0
+# 2                 1
+# 3                 3
+# 4                 6
+# => n*(n-1)/2
+# I
 class Solution:
     def numIdenticalPairs(self, nums: List[int]) -> int:
-        return sum(k * (k - 1) / 2 for k in collections.Counter(nums).values())
+        return sum(int(k * (k - 1) / 2) for k in collections.Counter(nums).values())
